// packages/api/src/types.ts
export type SophyDenomination =
  | 'wei'
  | 'ada'
  | 'babbage'
  | 'shannon'
  | 'szabo'
  | 'finney'
  | 'sophy'
  | 'ksophy'
  | 'msophy'
  | 'gsophy'
  | 'tsophy';